Fix signed-compare warning in wvdsp
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@1159 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ linuxstreams/wvdsp.cc
@@ -328,7 +328,7 @@ size_t WvDsp::do_uread(void *buf, size_t len)
     if (len < frag_size)
         log(WvLog::Warning, "reading less than frag size: %s/%s\n", len, frag_size);
 
-    int i, i2;
+    size_t i, i2;
     
     if (len > frag_size)
 	len = frag_size;
